several changes to db column types

+++ src/main/java/com/jimm0063/magi/debt/management/debtmanagementltesystem/domain/dto/AlmostCompletedDebtsDto.java
@@ -3,12 +3,14 @@ package com.jimm0063.magi.debt.management.debtmanagementltesystem.domain.dto;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Setter
 @Getter
 public class AlmostCompletedDebtsDto {
     private String code;
     private String description;
-    private Double monthlyPayment;
+    private BigDecimal monthlyPayment;
     private Integer currentInstallment;
     private Integer maxFinancingTerm;
 }
