+++ src/main/java/com/jimm0063/magi/debt/management/debtmanagementltesystem/domain/model/FixedExpense.java
@@ -3,12 +3,14 @@ package com.jimm0063.magi.debt.management.debtmanagementltesystem.domain.model;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Setter
 @Getter
 public class FixedExpense {
     private Integer id;
     private String name;
-    private Double monthlyCost;
+    private BigDecimal monthlyCost;
     private Integer paymentDay;
     private Boolean active;
     private DebtSysUser debtSysUser;
